Simplify commitgraph updater.
